requirements.txt
# FastAPI and Core Web Dependencies
fastapi>=0.115.11
uvicorn
jinja2
python-multipart

# Utility and Configuration
python-dotenv
requests

# Async and Database
asyncpg
sqlalchemy>=1.4.54
alembic>=1.15.1
psycopg[binary]

# Pydantic (for data validation)
pydantic[email,timezone]>=2.5.0

# Redis Task Queue
redis>=5.1.1
schedule

# Data Processing
unidecode
numpy>=1.26.4
pandas>=2.2.2
pendulum>=3.0.0
apache-airflow>=2.10.5

# Vector Search
qdrant-client>=1.13.0
fastembed
sentence-transformers>=3.4.1

# AI & ML from Google
transformers
google-genai>=1.5.0
google-cloud-translate>=3.15.3
google-cloud-texttospeech
langdetect>=1.0.9

# AI & ML from AWS
boto3

# PDF & Markdown Handling & Google Sheets
markdown
pypdl
gdown
oauth2client
gspread
pymupdf

# Media Processing
opencv-python-headless
moviepy